added jws.ts and removed key.ts
